requirements.txt
streamlit>=1.34
requests>=2.31
requests-oauthlib>=1.3
python-dotenv>=1.0
yfinance>=0.2
pandas>=2.0
pandas-datareader>=0.10
ta>=0.11
websocket-client>=1.8